% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-census-data.R
\name{hs_get_census_data}
\alias{hs_get_census_data}
\title{Get raw data from US census}
\usage{
hs_get_census_data(state = "AZ", year = 2022)
}
\arguments{
\item{state}{A two-letter abbreviation for a US state (case-insensitive).}

\item{year}{Year for which data are to be obtained, must be between 2009 and
year prior to present.}
}
\value{
A single, \pkg{sf}-formatted \code{data.frame} with geometries, and
columns for the five main variables.
}
\description{
Get raw data from US census
}
