% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ushardship-package.R
\docType{package}
\name{ushardship-package}
\alias{ushardship-package}
\alias{_PACKAGE}
\title{ushardship: Hardship Index from United States Census Data}
\description{
Hardship Index from United States Census Data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UrbanAnalyst/us-hardship-index}
  \item Report bugs at \url{https://github.com/UrbanAnalyst/us-hardship-index/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
