% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-census-data.R
\name{hs_get_census_data}
\alias{hs_get_census_data}
\title{Get raw data from US census}
\usage{
hs_get_census_data(state = "AZ", year = 2022, survey = "acs5")
}
\arguments{
\item{state}{A two-letter abbreviation for a US state (case-insensitive).}

\item{year}{Year for which data are to be obtained, must be between 2009 and
year prior to present.}

\item{survey}{Either "acs5" (the default) for five-year aggregated values,
or "acs1". Anuual availabilities differ for the different surveys. Data from
both surveys are available from 2010, for "acs5" generally to two years prior
to current year, and for "acs1" generally to one year prior.}
}
\value{
A single, \pkg{sf}-formatted \code{data.frame} with geometries, and
columns for the five main variables.
}
\description{
This returns the five primary variables required for the hardship index:
\itemize{
\item \code{occupancy}: Proportion of rooms with > 1 occupant
\item \code{poverty}: Proportion of households below poverty
\item \code{unemployment}: Proportion of unemployed adults (between 16 and
64)
\item \code{no_hs}: Proportion of population > 25 without a highschool
diploma
\item \code{prop_deps}: Proportion of population who may be considered
dependent; that is either under 18 or over 65
\item \code{income}: Per-capita income
}
}
